Changed to render with float precsion

+++ include/Components/AiComponent.h
@@ -3,7 +3,7 @@
 #include "Timer.h"
 struct Ai
 {
-    int m_moveDirectionX;
-    int m_moveDirectionY;
+    float m_moveDirectionX;
+    float m_moveDirectionY;
     Timer m_clock;
 };
